% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{addData,Collection-method}
\alias{addData,Collection-method}
\title{This is the basic training endpoint. Given a piece of text and a score, either categorical
or numeric, this endpoint will train a new model given the additional piece of information.}
\usage{
\S4method{addData}{Collection}(collection_object, data, version = NULL,
  domain = NULL, api_key = FALSE, ...)
}
\arguments{
\item{data}{the text and collection/score associated with it. The length of the text (string) should ideally
be longer than 100 characters and contain at least 10 words. While the API will support
shorter text, you will find that the accuracy of results improves significantly with longer
examples. For an additional fee, this end point will support image input as well. The collection/score
can be a string or float. This is the variable associated with the text. This can either be categorical
(the tag associated with the post) or numeric (the number of Facebook shares the post
received). However it can only be one or another within a given label.}

\item{version}{for api version}

\item{api_key}{your personal indico API key}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List with available collections
}
\description{
This is the basic training endpoint. Given a piece of text and a score, either categorical
or numeric, this endpoint will train a new model given the additional piece of information.
}
\examples{
collection <- Collection(name='example')
test_data <- list(list("I love my friends!", "extrovert"),
                  list("I love to be alone", "introvert"),
                  list("I have mixed feelings on people", "ambivert"))
addData(collection, test_data)
}
