% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{analyze_image}
\alias{analyze_image}
\title{Returns multiple image API results in a single request / response}
\usage{
analyze_image(img, apis = c("facial_features", "fer", "image_features",
  "facial_localization", "content_filtering"), version = NULL, ...)
}
\arguments{
\item{img}{image data}

\item{apis}{vector / array of API names as strings}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
map of apis as keys to their results as values
}
\description{
Given input text and a vector of API names, returns aggregated results of the apis.
}
\examples{
## Example 1
img <- "../tests/testthat/image/image.png"
emotion <- analyze_image(img, apis=c("fer"))

}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{multi}

