% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{analyze_text}
\alias{analyze_text}
\title{Returns multiple text API results in a single request / response}
\usage{
analyze_text(text, apis = c("sentiment", "text_tags", "political", "language",
  "keywords", "twitter_engagement", "named_entities"), api_key = FALSE,
  cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{apis}{vector / array of API names as strings}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
map of apis as keys to their results as values
}
\description{
Given input text and a vector of API names, returns aggregated results of the apis.
}
\examples{
results <- analyze_text("Thanks everyone for the birthday wishes!!
                      It was a crazy few days ><", apis=c("sentiment", "political"))
results
cat(sprintf("This text has \%s tonality",
             ifelse(results[["sentiment"]] > 0.5, "positive", "negative")))

}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{multi}

