% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/content_filtering.R
\name{batch_content_filtering}
\alias{batch_content_filtering}
\title{Detects NSFW content in a list of mages}
\usage{
batch_content_filtering(imgs, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{imgs}{image data}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List of Floats with probabilities of NSFW
}
\description{
Given a list of input images,  returns a list of float probabilities between 0 and 1 that content is NSFW
}
\details{
* Input can be either grayscale or rgb color and should either be a numpy array or nested list format.

* Input data should be either uint8 0-255 range values or floating point between 0 and 1.

* Large images (i.e. 1024x768+) are much bigger than needed, resizing will be done internally to 64x64 if needed.

* For ideal performance, images should be square aspect ratio but non-square aspect ratios are supported as well.
}
\examples{
img_list = list()
img_list[[1]] = matrix(runif(64*64, 0, 1), nrow = 64)
probabilities <- batch_content_filtering(img_list)
}
\seealso{
\code{\link{face_emotion}}, \code{\link{facial_features}}
}
\keyword{API}
\keyword{content}
\keyword{filtering}
\keyword{image}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}

