% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/face_emotion.R
\name{batch_face_emotion}
\alias{batch_face_emotion}
\title{Detects face emotion}
\usage{
batch_face_emotion(imgs, api_key = FALSE, cloud = FALSE)
}
\arguments{
\item{imgs}{image data}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List of lists with face emotions probability pairs
}
\description{
Given a list of grayscale input images of faces, returns a list of probability distributions over emotional state.
Input should be a list of two-dimensional structures (like data.frame or matrix), resizing will be attempted internally but for best
performance, images should be already sized at 48x48 pixels.
}
\examples{
## Example 1
img_list = list()
img_list[[1]] <- matrix(runif(48*48, 0, 1), nrow = 48)
emotion <- batch_face_emotion(img_list)

most.possible <- sort(unlist(emotion[[1]]), decreasing = TRUE)[1:2]
cat(sprintf("Detected '\%s' emotion with probability \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is '\%s' emotion with probability \%0.4f.",
            names(most.possible)[2], most.possible[2]))
}
\seealso{
\code{\link{facial_features}}, \code{\link{image_features}}
}
\keyword{API}
\keyword{emotions}
\keyword{face}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

