% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facial_features.R
\name{batch_facial_features}
\alias{batch_facial_features}
\title{Detects facial features}
\usage{
batch_facial_features(imgs, auth = FALSE, cloud = FALSE)
}
\arguments{
\item{imgs}{image data}

\item{auth}{Username and password for HTTP Basic Auth}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List of lists with facial features
}
\description{
Given a list of grayscale input images of a face, returns a 48 dimensional feature vector explaining that face.
Useful as a form of feature engineering for face oriented tasks.
Input should be in a list of list of lists format, resizing will be attempted internally but for best
performance, images should be already sized at 48x48 pixels.
}
\examples{
img_list = list()
img_list[[1]] = matrix(runif(48*48, 0, 1), nrow = 48)
features <- facial_features(img_list)
length(features)
length(features[[1]])
}
\seealso{
\code{\link{face_emotion}}, \code{\link{image_features}}
}
\keyword{API}
\keyword{face}
\keyword{features}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

