% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/image_features.R
\name{batch_image_features}
\alias{batch_image_features}
\title{Detects image features for a list of images}
\usage{
batch_image_features(imgs, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{imgs}{image data}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List of lists with image features
}
\description{
Given a list of input images, returns a list of 2048 dimensional sparse feature vectors explaining that image.
Useful as a form of feature engineering for image oriented tasks.
}
\details{
* Input can be either grayscale or rgb color and should either be a numpy array or nested list format.

* Input data should be either uint8 0-255 range values or floating point between 0 and 1.

* Large images (i.e. 1024x768+) are much bigger than needed, resizing will be done internally to 64x64 if needed.

* For ideal performance, images should be square aspect ratio but non-square aspect ratios are supported as well.
}
\examples{
img_list = list()
img_list[[1]] = matrix(runif(64*64, 0, 1), nrow = 64)
features <- batch_image_features(img_list)

length(features)
length(features[[1]])
min(unlist(features[[1]]))
max(unlist(features[[1]]))
sum(unlist(features[[1]]))
}
\seealso{
\code{\link{face_emotion}}, \code{\link{facial_features}}
}
\keyword{API}
\keyword{features}
\keyword{image}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

