% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keywords.R
\name{batch_keywords}
\alias{batch_keywords}
\title{Tags text with likely categories}
\usage{
batch_keywords(text, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{text}{text for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List of lists with text tag probability pairs
}
\description{
Given a list of input documents, returns a list of probability distributions
over 111 possible topics
}
\examples{
keywords <- batch_keywords(c("Monday: Delightful with mostly sunny skies",
                                 "Highs in the low 70s."))
keywords
most.possible <- sort(unlist(keywords[[1]]), decreasing = TRUE)[1:2]
cat(sprintf("Detected keyword \%s with a score of \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is \%s with a score of \%0.4f.",
            names(most.possible)[2], most.possible[2]))
}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{classification}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{tagging}

