% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/language.R
\name{batch_language}
\alias{batch_language}
\title{Detects language of the given documents}
\usage{
batch_language(text, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{text}{array of documents for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List of lists with language probability pairs
}
\description{
Given a list of documents, returns a list of probability distributions over 33 possible
languages
}
\examples{
languages <- language(c("Monday: Delightful with mostly sunny skies.
                         Highs in the low 70s."))
languages
most.possible <- sort(unlist(languages[[1]]), decreasing = TRUE)[1:2]
cat(sprintf("Detected \%s language with probability \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is \%s with probability \%0.4f.",
            names(most.possible)[2], most.possible[2]))
}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{detection}
\keyword{indico.io}
\keyword{language}
\keyword{learning}
\keyword{machine}

