% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/named_entities.R
\name{batch_named_entities}
\alias{batch_named_entities}
\title{Named entities with likely categories}
\usage{
batch_named_entities(text, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{text}{text for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List of lists entity category-probability pairs
}
\description{
Given input text, returns named entities and the probabilities of what category they are in
}
\examples{
text_list <- list()
text_list[[1]] <- "London Undergroud's boss Mike Brown warned of the strike..."
entities <- batch_named_entities(text_list)
}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{classification}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{tagging}

