% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/political.R
\name{batch_political}
\alias{batch_political}
\title{Returns a list of probability distributions over the political alignments of the speakers}
\usage{
batch_political(text, auth = FALSE, cloud = FALSE)
}
\arguments{
\item{text}{array of documents for analysis}

\item{auth}{username and password for HTTP Basic Auth}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List of lists with party probability pairs
}
\description{
Given input text, returns a probability distribution over the political
alignment of the speaker.
}
\examples{
text <- c("I am so proud to stand here today
           as Prime Minister of four nations
           in one United Kingdom.")
affilation <- batch_political()
affilation
most.like <- names(affilation[which.max(unlist(affilation[[1]]))])
least.like <- names(affilation[which.min(unlist(affilation[[1]]))])
cat(sprintf("This text is most like \%s and least like \%s",
            most.like, least.like))
}
\seealso{
\code{\link{sentiment}}, \code{\link{language}}
}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{political}
\keyword{sentiment}

