% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi.R
\name{batch_predict_image}
\alias{batch_predict_image}
\title{Returns multiple batched image API results in a single request / response}
\usage{
batch_predict_image(imgs, apis = c("facial_features", "face_emotion",
  "image_features"), api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{imgs}{image data}

\item{apis}{vector / array of API names as strings}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
map of apis as keys to their results as values
}
\description{
Given a list of grayscale input images of faces, returns a map of returns aggregated results of the apis.
}
\examples{
## Example 1
img_list = list()
img_list[[1]] <- matrix(runif(48*48, 0, 1), nrow = 48)
emotion <- batch_predict_image(img_list)
}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{multi}

