% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentiment.R
\name{batch_sentiment}
\alias{batch_sentiment}
\title{Returns a list of scalar sentiment estimates}
\usage{
batch_sentiment(text, api_key = FALSE, cloud = FALSE)
}
\arguments{
\item{text}{array of documents for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List of numerical values roughly in the range from 0 to 1 with 0.5 indicating neutral sentiment.
For reference, 0 suggests very negative sentiment and 1 suggests very positive sentiment.
}
\description{
Given a list of input documents, returns a list of scalar sentiment estimates
}
\examples{
text_list <- list()
text_list[[1]] <- "Thanks everyone for the birthday wishes!!
                              It was a crazy few days ><"
emotion <- batch_sentiment(text_list)
emotion
cat(sprintf("This text has \%s tonality",
             ifelse(emotion[[1]] > 0.5, "positive", "negative")))
}
\seealso{
\code{\link{political}}, \code{\link{language}}
}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{sentiment}

