% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text_tags.R
\name{batch_text_tags}
\alias{batch_text_tags}
\title{Tags text with likely categories}
\usage{
batch_text_tags(text, auth = FALSE, cloud = FALSE)
}
\arguments{
\item{text}{text for analysis}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List of lists with text tag probability pairs
}
\description{
Given a list of input documents, returns a list of probability distributions
over 111 possible topics
}
\examples{
categories <- batch_text_tags(c("Monday: Delightful with mostly sunny skies.
                                 Highs in the low 70s."))
categories
most.possible <- sort(unlist(categories[[1]]), decreasing = TRUE)[1:2]
cat(sprintf("Detected category \%s with probability \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is \%s with probability \%0.4f.",
            names(most.possible)[2], most.possible[2]))
}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{classification}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{tagging}

