% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_engagement.R
\name{batch_twitter_engagement}
\alias{batch_twitter_engagement}
\title{Returns a list of floats indicating engagement scores for the list of text}
\usage{
batch_twitter_engagement(text, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{text}{array of documents for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List of lists with party probability pairs
}
\description{
Given input text, returns a probability distribution over the political
alignment of the speaker.
}
\examples{
text_list <- list()
text_list[[1]] <- "I am so proud to stand here today
           as Prime Minister of four nations
           in one United Kingdom."
engagements <- batch_twitter_engagement(text_list)
cat(sprintf("This text has twitter engagement of \%f",
            engagements[[1]]))
}
\seealso{
\code{\link{sentiment}}, \code{\link{language}}
}
\keyword{API}
\keyword{analysis}
\keyword{engagement}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{twitter}

