% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{clear,Collection-method}
\alias{clear,Collection-method}
\title{This is an API made to remove all of the data associated from a given colletion. If there's been a data
corruption issue, or a large amount of incorrect data has been fed into the API it is often difficult
to correct. This allows you to clear a colletion and start from scratch. Use with caution! This is not
reversible.}
\usage{
\S4method{clear}{Collection}(collection_object, data, api_key = FALSE,
  cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}
}
\value{
List with available collections
}
\description{
This is an API made to remove all of the data associated from a given colletion. If there's been a data
corruption issue, or a large amount of incorrect data has been fed into the API it is often difficult
to correct. This allows you to clear a colletion and start from scratch. Use with caution! This is not
reversible.
}
\examples{
collection <- Collection(name='example')
clear(collection)
}

