% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_filtering.R
\name{content_filtering}
\alias{content_filtering}
\title{Detects NSFW content in an image}
\usage{
content_filtering(img, version = NULL, ...)
}
\arguments{
\item{img}{image data}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
Float with probability of NSFW
}
\description{
Given an input image, returns a float probability between 0 and 1 that content is NSFW
}
\details{
* Input can be either grayscale or rgb color and should either be a numpy array or nested list format.

* Input data should be either uint8 0-255 range values or floating point between 0 and 1.

* Large images (i.e. 1024x768+) are much bigger than needed, resizing will be done internally to 64x64 if needed.

* For ideal performance, images should be square aspect ratio but non-square aspect ratios are supported as well.
}
\examples{
img <- "../tests/testthat/image/image.png"
probabilities <- content_filtering(img)

}
\seealso{
\code{\link{face_emotion}}, \code{\link{facial_features}}
}
\keyword{API}
\keyword{content}
\keyword{filtering}
\keyword{image}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}

