% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{convert_results}
\alias{convert_results}
\title{Converts response list to one that uses user inputted api names}
\usage{
convert_results(results, apis)
}
\arguments{
\item{results}{response list}

\item{apis}{vector of provided string api names}
}
\value{
converted_results converted response list
}
\description{
Given response list and list of user input apis, return properly named response
}

