% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{deauthorize,Collection-method}
\alias{deauthorize,Collection-method}
\title{Remove a given user's access to the current collection}
\usage{
\S4method{deauthorize}{Collection}(collection_object, email, ...)
}
\arguments{
\item{email}{the email of the user you'd like to remove access from}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}
}
\value{
Boolean that indicates success or failure
}
\description{
Remove a given user's access to the current collection
}
\examples{
collection <- Collection(name = 'example')
success = deauthorize(collection, 'contact@indico.io')
}
