% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{deregister,Collection-method}
\alias{deregister,Collection-method}
\title{Deregister the current collection so that the collection is no longer shared with other users.}
\usage{
\S4method{deregister}{Collection}(collection_object, ...)
}
\arguments{
\item{collection_object}{the collection object for this model}

\item{...}{additional arguments to passed to request}
}
\value{
Boolean that indicates success or failure
}
\description{
Deregister the current collection so that the collection is no longer shared with other users.
}
\examples{
collection <- Collection(name = 'example')
success = deregister(collection)
}
