% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion.R
\name{emotion}
\alias{emotion}
\title{Detects emotion of the text}
\usage{
emotion(text, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with emotion probability pairs
}
\description{
Given input text, returns a probability distribution over 33 possible
languages of what language the text was written in.
}
\examples{
emotions <- emotion("Monday: Delightful with mostly sunny skies. Highs in the low 70s.")
most.possible <- sort(unlist(emotions), decreasing = TRUE)[1:2]
cat(sprintf("Detected \%s emotion with probability \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is \%s with probability \%0.4f.",
            names(most.possible)[2], most.possible[2]))

}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{detection}
\keyword{emotion}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}

