% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_features.R
\name{facial_features}
\alias{facial_features}
\title{Detects facial features}
\usage{
facial_features(img, version = NULL, ...)
}
\arguments{
\item{img}{image data}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with face features
}
\description{
Given an grayscale input image of a face, returns a 48 dimensional feature vector explaining that face.
Useful as a form of feature engineering for face oriented tasks.
Input should be in a list of list format, resizing will be attempted internally but for best
performance, images should be already sized at 48x48 pixels.
}
\examples{
img <- "../tests/testthat/image/image.png"
features <- facial_features(img)
length(features)

}
\seealso{
\code{\link{face_emotion}}, \code{\link{image_features}}
}
\keyword{API}
\keyword{face}
\keyword{features}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}
