% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facial_localization.R
\name{facial_localization}
\alias{facial_localization}
\title{Detects faces}
\usage{
facial_localization(img, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{img}{image data}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List with face locations
}
\description{
Given a grayscale input image, returns a possible locations for faces
Input should be a two-dimensional structure (like data.frame or matrix), resizing will be attempted internally but for best
performance, images should be already sized at 48x48 pixels.
}
\examples{
## Example 1
img <- "../tests/testthat/image/image.png"
emotion <- facial_localization(img)
}
\seealso{
\code{\link{facial_features}}, \code{\link{image_features}}
}
\keyword{API}
\keyword{emotions}
\keyword{face}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

