% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi.R
\name{filter_map}
\alias{filter_map}
\title{Checks if api is allowed and maps it to the REST Api name}
\usage{
filter_map(input, allowed)
}
\arguments{
\item{input}{vector of string api names}

\item{allowed}{vector of acceptable string api names}
}
\value{
vector of converted api names
}
\description{
Given vector of input apis and vector of accepted APIs, returns vector of mapped apis
}

