% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_handler.R
\name{format_images}
\alias{format_images}
\title{Returns a list of `data.frame`s given a list of input images}
\usage{
format_images(imgs, size, min_axis = FALSE)
}
\arguments{
\item{imgs}{List of images to convert to a `string`s}

\item{size}{Size of image to resize to}

\item{min_axis}{Whether or not to keep aspect ratio}
}
\value{
`String`s constructed from list of images as base64 encoded
}
\description{
Given a list of input images, returns a list of `string`s
}

