% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_recognition.R
\name{image_recognition}
\alias{image_recognition}
\title{Detects image recognition}
\usage{
image_recognition(img, api_key = FALSE, cloud = FALSE, version = NULL,
  ...)
}
\arguments{
\item{img}{image data}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with image recognition
}
\description{
Given an input image, returns a series of image classifications and associated scores
}
\details{
* Input can be either grayscale or rgb color and should either be a numpy array or nested list format.

* Input data should be either uint8 0-255 range values or floating point between 0 and 1.

* Large images (i.e. 1024x768+) are much bigger than needed, resizing will be done internally to 64x64 if needed.

* For ideal performance, images should be square aspect ratio but non-square aspect ratios are supported as well.
}
\examples{
img <- "../tests/testthat/image/image.png"
features <- image_recognition(img)

length(features)
min(unlist(features))
max(unlist(features))
sum(unlist(features))

}
\seealso{
\code{\link{face_emotion}}, \code{\link{facial_features}}
}
\keyword{API}
\keyword{image}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

