% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{info,Collection-method}
\alias{info,Collection-method}
\title{Return the current state of the model associated with a given collection}
\usage{
\S4method{info}{Collection}(collection_object, version = NULL, ...)
}
\arguments{
\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List with available collections
}
\description{
Return the current state of the model associated with a given collection
}
\examples{
collection <- Collection(name='example')
status = info(collection)
cat(sprintf("This collection is a \%s model trained on \%s data with \%i examples",
            status[["model_type"]],
            status[["input_type"]],
            status[["number_of_examples"]]))
}
