% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi.R
\name{intersections}
\alias{intersections}
\title{Returns strength of correlation between API results}
\usage{
intersections(data, apis = FALSE, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{data}{input data}

\item{apis}{vector of provided string api names}
}
\value{
nested associated array of correlation strengths
}
\description{
Given an array of data and a vector of apis, returns a nested associate array of correlation strengths
}

