% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{keywords}
\alias{keywords}
\title{Tags text with likely keywords}
\usage{
keywords(text, version = 2, language = "english", ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{language}{A string}

\item{...}{additional arguments to passed to request}
}
\value{
List with keyword score pairs
}
\description{
Given input text, returns a series of keywords and associated scores
}
\examples{
keywords <- keywords("Monday: Delightful with mostly sunny skies.
                   Highs in the low 70s.")
keywords
most.possible <- sort(unlist(keywords), decreasing = TRUE)[1:2]
cat(sprintf("Detected keyword \%s with a score of \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is \%s with a score \%0.4f.",
            names(most.possible)[2], most.possible[2]))

}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}, \code{\link{text_tags}}
}
\keyword{API}
\keyword{classification}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{tagging}
