% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_handler.R
\name{make_request}
\alias{make_request}
\title{Returns a response from the indico API endpoint}
\usage{
make_request(data, api, api_key = FALSE, cloud = FALSE, version = NULL,
  ...)
}
\arguments{
\item{data}{data for analysis}

\item{api}{api to call}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{batch}{to make a batched request}

\item{batch}{send call to batch endpoint}
}
\value{
error or response extracted from the indico API response
}
\description{
Given input data, return the response from the indico API endpoint specified
}
\keyword{API}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}

