% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_entities.R
\name{named_entities}
\alias{named_entities}
\title{Named entities with likely categories}
\usage{
named_entities(text, api_key = FALSE, cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with a list with entitiy category list pairs, the categor list is of category probability pairs
}
\description{
Given input text, returns named entities and the probabilities of what category they are in
}
\examples{
entities <- named_entities("London Undergroud's boss Mike Brown warned of the strike...")
entities
}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{classification}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{tagging}

