% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{organizations}
\alias{organizations}
\title{Finds the names of organizations in the input text}
\usage{
organizations(text, api_key = FALSE, cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
array of dicts, each with 'text', 'confidence', and 'position' fields
}
\description{
Finds the names of organizations in the input text
}
\examples{
result <- organizations("A year ago, the New York Times published confidential comments about ISIS' ideology by Major General Michael K. Nagata, then U.S. Special Operations commander in the Middle East.")
cat(sprintf("\%s is mentioned in the input text" \% result[0]['text']))

}
\seealso{
\code{\link{places}}, \code{\link{organizations}}
}

