% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personas.R
\name{personas}
\alias{personas}
\title{Predicts the likelyhood the author of the text is 16 different personas}
\usage{
personas(text, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
dict with likelihood the author would be catigorized as each myers briggs axis,
and the most likley MB type
}
\description{
Predicts the likelyhood the author of the text is 16 different personas
}
\examples{
personas <- personas("Monday: Delightful with mostly sunny skies.
                   Highs in the low 70s.")
most.possible <- sort(unlist(personas), decreasing = TRUE)[1:2]
cat(sprintf("Detected \%s persona with probability \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))

}
\seealso{
\code{\link{personality}}, \code{\link{sentiment}}, \code{\link{text_tags}}
}
