% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{predict,Collection-method}
\alias{predict,Collection-method}
\title{This is the prediction endpoint. This will be the primary interaction point for all predictive
 analysis.}
\usage{
\S4method{predict}{Collection}(collection_object, data, api_key = FALSE,
  cloud = FALSE, version = NULL, domain = NULL, ...)
}
\arguments{
\item{data}{the text example being provided to the API. As a general rule, the data should be as
similar to the examples given to the train function (above) as possible. Because language
in different domains is used very differently the accuracy will generally drop as the
difference between this text and the training text increases. Base64 encoded image data, image urls, and
text content are all valid.}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{domain}{String: This is an identifier that helps determine the appropriate techniques for indico
to use behind the scenes to train your model.  One of {"standard", "topics"}}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model
the response format for the given label will match the format of the training examples}
}
\value{
List with available collections
}
\description{
This is the prediction endpoint. This will be the primary interaction point for all predictive
 analysis.
}
\examples{
collection <- Collection(name='example')
test_data <- list(list("I love my friends!", "extrovert"),
                  list("I love to be alone", "introvert"),
                  list("I have mixed feelings on people", "ambivert"))
addData(collection, test_data)
train(collection)
wait(collection)
res = train("I love my friends!")
cat(sprintf("The likelihood the author was an extrovert is \\\%0.4f.",
            res[["extrovert"]])
}

