% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi.R
\name{predict_image}
\alias{predict_image}
\title{Returns multiple image API results in a single request / response}
\usage{
predict_image(img, apis = c("facial_features", "face_emotion",
  "image_features"), api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{img}{image data}

\item{apis}{vector / array of API names as strings}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
map of apis as keys to their results as values
}
\description{
Given input text and a vector of API names, returns aggregated results of the apis.
}
\examples{
## Example 1
img <- matrix(runif(48*48, 0, 1), nrow = 48)
emotion <- predict_image(img, apis=c("face_emotion"))
}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{multi}

