% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi.R
\name{predict_text}
\alias{predict_text}
\title{Returns multiple text API results in a single request / response}
\usage{
predict_text(text, apis = c("sentiment", "text_tags", "political",
  "language"), api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{text}{text for analysis}

\item{apis}{vector / array of API names as strings}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
map of apis as keys to their results as values
}
\description{
Given input text and a vector of API names, returns aggregated results of the apis.
}
\examples{
results <- predict_text("Thanks everyone for the birthday wishes!!
                      It was a crazy few days ><", apis=c("sentiment", "political"))
results
cat(sprintf("This text has \%s tonality",
             ifelse(results[["sentiment"]] > 0.5, "positive", "negative")))
}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{multi}

