% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{remove_example,Collection-method}
\alias{remove_example,Collection-method}
\title{This is an API made to remove a single instance of training data. This is useful in cases where a
 single instance of content has been modified, but the remaining examples remain valid. For
 example, if a piece of content has been retagged.}
\usage{
\S4method{remove_example}{Collection}(collection_object, data, ...)
}
\arguments{
\item{collection_object}{the collection object for this model}

\item{data}{The exact text you wish to remove from the given collection. If the string
provided does not match a known piece of text then this will fail. Again, this is required if
an id is not provided, and vice-versa.}

\item{...}{additional arguments to passed to request}
}
\value{
List with available collections
}
\description{
This is an API made to remove a single instance of training data. This is useful in cases where a
 single instance of content has been modified, but the remaining examples remain valid. For
 example, if a piece of content has been retagged.
}
\examples{
collection <- Collection(name = 'example')
test_data <- list(list("I love my friends!", "extrovert"),
                  list("I love to be alone", "introvert"),
                  list("I have mixed feelings on people", "ambivert"))
addData(collection, test_data)
remove_example(collection, test_data[[1]][[1]])
}
