% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{text_features}
\alias{text_features}
\title{Transforms text inputs into numerical features}
\usage{
text_features(text, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with image features
}
\description{
Transforms text inputs into numerical features
}
\examples{
features <- text_features("Some text for testing this awesome text features api.")

length(features) # 300

}
\seealso{
\code{\link{image_features}}, \code{\link{relevance}}
}
\keyword{API}
\keyword{features}
\keyword{image}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

