% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tags.R
\name{text_tags}
\alias{text_tags}
\title{Tags text with likely categories}
\usage{
text_tags(text, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with text tag probability pairs
}
\description{
Given input text, returns a probability distribution over 111 possible topics
}
\examples{
categories <- text_tags("Monday: Delightful with mostly sunny skies.
                   Highs in the low 70s.")
categories
most.possible <- sort(unlist(categories), decreasing = TRUE)[1:2]
cat(sprintf("Detected category \%s with probability \%0.4f.\\n",
            names(most.possible)[1], most.possible[1]))
cat(sprintf("Next possible is \%s with probability \%0.4f.",
            names(most.possible)[2], most.possible[2]))

}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}
}
\keyword{API}
\keyword{classification}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{tagging}

