% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{train,Collection-method}
\alias{train,Collection-method}
\title{This is the basic training endpoint. Given an existing dataset this endpoint will train a model.}
\usage{
\S4method{train}{Collection}(collection_object, api_key = FALSE,
  cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}
}
\value{
List with available collections
}
\description{
This is the basic training endpoint. Given an existing dataset this endpoint will train a model.
}
\examples{
collection <- Collection(name='example')
train(collection)
}

