% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{authorize,Collection-method}
\alias{authorize,Collection-method}
\title{Authorize a given user to access the current collection}
\usage{
\S4method{authorize}{Collection}(collection_object, email,
  permission_type = "read", ...)
}
\arguments{
\item{email}{the email of the user you'd like to give access to}

\item{permission_type}{read / write -- the type of permission to give the authorized user}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}
}
\value{
Boolean that indicates success or failure
}
\description{
Authorize a given user to access the current collection
}
\examples{
collection <- Collection(name = 'example')
success = authorize(collection, 'contact@indico.io')
}
