% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collections}
\alias{collections}
\title{This is a status report endpoint. It is used to get the status on all of the collections currently trained, as
well as some basic statistics on their accuracies. See docs for more information}
\usage{
collections(api_key = FALSE, cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
List with available collections
}
\description{
This is a status report endpoint. It is used to get the status on all of the collections currently trained, as
well as some basic statistics on their accuracies. See docs for more information
}
\examples{
collections <- collections()

cat(sprintf("There are currently \%i collections",
            length(collections))
}

