% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facial_features.R
\name{facial_features}
\alias{facial_features}
\title{Detects facial features}
\usage{
facial_features(img, api_key = FALSE, cloud = FALSE, ...)
}
\arguments{
\item{img}{image data}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{...}{additional arguments to passed to request}
}
\value{
List with face features
}
\description{
Given an grayscale input image of a face, returns a 48 dimensional feature vector explaining that face.
Useful as a form of feature engineering for face oriented tasks.
Input should be in a list of list format, resizing will be attempted internally but for best
performance, images should be already sized at 48x48 pixels.
}
\examples{
img <- matrix(runif(48*48, 0, 1), nrow = 48)
features <- facial_features(img)
length(features)
}
\seealso{
\code{\link{face_emotion}}, \code{\link{image_features}}
}
\keyword{API}
\keyword{face}
\keyword{features}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{recognition}

