% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_handler.R
\name{format_image}
\alias{format_image}
\title{Returns a response from the indico API endpoint}
\usage{
format_image(img, size, min_axis = FALSE)
}
\arguments{
\item{img}{Image to convert to a data.frame}

\item{size}{Size of image to resize to}

\item{min_axis}{Whether or not to keep aspect ratio '}
}
\value{
string base64 encoding of resized image
}
\description{
Given an input image, returns a data.frame obj
}

