% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api_handler.R
\name{format_images}
\alias{format_images}
\title{Returns a list of `data.frame`s given a list of input images}
\usage{
format_images(imgs)
}
\arguments{
\item{imgs}{List of images to convert to a `data.frame`s}
}
\value{
`data.frame`s constructed from list of images
}
\description{
Given a list of input images, returns a list of `data.frame`s
}

