% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{info,Collection-method}
\alias{info,Collection-method}
\title{Return the current state of the model associated with a given collection}
\usage{
\S4method{info}{Collection}(collection_object, api_key = FALSE,
  cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}
}
\value{
List with available collections
}
\description{
Return the current state of the model associated with a given collection
}
\examples{
collection <- Collection(name='example')
status = info(collection)
cat(sprintf("This collection is a \%s model trained on \%s data with \%i examples",
            status[[model_type]],
            status[[input_type]],
            status[[number_of_examples]])
}

