% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api_handler.R
\name{make_request}
\alias{make_request}
\title{Returns a response from the indico API endpoint}
\usage{
make_request(data, api, auth = FALSE, cloud = FALSE, batch = FALSE)
}
\arguments{
\item{data}{data for analysis}

\item{api}{api to call}

\item{auth}{username and password for HTTP Basic Auth}

\item{cloud}{subdomain for indico private cloud}

\item{batch}{send call to batch endpoint}
}
\value{
error or response extracted from the indico API response
}
\description{
Given input data, return the response from the indico API endpoint specified
}
\keyword{API}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}

