% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people}
\alias{people}
\title{Finds the names and titles of people in the input text}
\usage{
people(text, api_key = FALSE, cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
array of dicts, each with 'text', 'confidence', and 'position' fields
}
\description{
Finds the names and titles of people in the input text
}
\examples{
result <- people("Barack Obama is scheduled to give a talk next Saturday at the White House.")
cat(sprintf("\%s is mentioned in the input text", result[0]['text']))

}
\seealso{
\code{\link{places}}, \code{\link{organizations}}
}

