% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personality.R
\name{personality}
\alias{personality}
\title{Predicts a score between 0 and 1 for 4 personality traits for the author of the texts}
\usage{
personality(text, api_key = FALSE, cloud = FALSE, version = NULL, ...)
}
\arguments{
\item{text}{text for analysis}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
dict with likelihood the author would be catigorized as each myers briggs axis,
and the most likley MB type
}
\description{
Predicts a score between 0 and 1 for 4 personality traits for the author of the texts.
The personality scales are "openness", "extraversion", "conscientiousness", and "agreeableness"
}
\examples{
personality <- personality("Monday: Delightful with mostly sunny skies.
                   Highs in the low 70s.")
cat(sprintf("The author of text is \%s",
             ifelse(personality['extraversion'] < .5, "introverted", "extraverted")))

}
\seealso{
\code{\link{political}}, \code{\link{sentiment}}, \code{\link{text_tags}}
}

