% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{places}
\alias{places}
\title{Finds the names of places in the input text}
\usage{
places(text, version = 2, ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}
}
\value{
array of dicts, each with 'text', 'confidence', and 'position' fields
}
\description{
Finds the names of places in the input text
}
\seealso{
\code{\link{people}}, \code{\link{organizations}}
}

