% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political.R
\name{political}
\alias{political}
\title{Returns a probability distribution over the political alignment of the speaker}
\usage{
political(text, version = 2, ...)
}
\arguments{
\item{text}{text for analysis}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{api_key}{your personal indico API key}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List with party probability pairs
}
\description{
Given input text, returns a probability distribution over the political
alignment of the speaker.
}
\examples{
affilation <- political("I am so proud to stand here today
                         as Prime Minister of four nations
                         in one United Kingdom.")
affilation
most.like <- names(affilation[which.max(unlist(affilation))])
least.like <- names(affilation[which.min(unlist(affilation))])
cat(sprintf("This text is most like \%s and least like \%s",
            most.like, least.like))

}
\seealso{
\code{\link{sentiment}}, \code{\link{language}}
}
\keyword{API}
\keyword{analysis}
\keyword{indico.io}
\keyword{learning}
\keyword{machine}
\keyword{political}
\keyword{sentiment}
