% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{predict,Collection-method}
\alias{predict,Collection-method}
\title{This is the prediction endpoint. This will be the primary interaction point for all predictive
 analysis.}
\usage{
\S4method{predict}{Collection}(collection_object, data, ...)
}
\arguments{
\item{collection_object}{the collection object for this model
the response format for the given label will match the format of the training examples}

\item{data}{the text example being provided to the API. As a general rule, the data should be as
similar to the examples given to the train function (above) as possible. Because language
in different domains is used very differently the accuracy will generally drop as the
difference between this text and the training text increases. Base64 encoded image data, image urls, and
text content are all valid.}

\item{...}{additional arguments to passed to request}
}
\value{
List with available collections
}
\description{
This is the prediction endpoint. This will be the primary interaction point for all predictive
 analysis.
}
\examples{
collection <- Collection(name = 'example')
test_data <- list(list("I love my friends!", "extrovert"),
                  list("I love to be alone", "introvert"),
                  list("I have mixed feelings on people", "ambivert"))
addData(collection, test_data)
train(collection)
wait(collection)
res <- predict(collection, "I love my friends!")
cat(sprintf("The likelihood the author was an extrovert is \\\%0.4f.",
            res[["extrovert"]]))
}
