% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{register,Collection-method}
\alias{register,Collection-method}
\title{Register the current collection so that the collection can be shared with other users.}
\usage{
\S4method{register}{Collection}(collection_object, version = NULL, ...)
}
\arguments{
\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}

\item{cloud}{subdomain for indico private cloud}
}
\value{
Boolean that indicates success or failure
}
\description{
Register the current collection so that the collection can be shared with other users.
}
\examples{
collection <- Collection(name='example')
success = register(collection)
}
