% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{rename,Collection-method}
\alias{rename,Collection-method}
\title{Rename the current model to a new name.}
\usage{
\S4method{rename}{Collection}(collection_object, name, version = NULL, ...)
}
\arguments{
\item{name}{the new name to be used to access this collection}

\item{version}{for api version}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}

\item{cloud}{subdomain for indico private cloud}
}
\value{
Boolean that indicates success or failure
}
\description{
Rename the current model to a new name.
}
\examples{
collection <- Collection(name='example')
success = rename(collection, 'new-example')
}
