% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_handler.R
\name{request_url}
\alias{request_url}
\title{Returns the url for the proper indico API}
\usage{
request_url(cloud, api, batch, api_key, version = NULL, apis = NULL, ...)
}
\arguments{
\item{cloud}{(optional) indico subdomain for private cloud}

\item{api}{name of API}

\item{batch}{(logical) does the request contain more than one example?}

\item{api_key}{your personal indico API key}

\item{version}{for api version}

\item{apis}{possible list of apis url paramater for multi qpi requests}

\item{...}{additional possible arguments to passed as url parameters}
}
\value{
url for API request
}
\description{
Produces the proper url to query with a given request
}

