% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{train,Collection-method}
\alias{train,Collection-method}
\title{This is the basic training endpoint. Given an existing dataset this endpoint will train a model.}
\usage{
\S4method{train}{Collection}(collection_object, ...)
}
\arguments{
\item{collection_object}{the collection object for this model}

\item{...}{additional arguments to passed to request}
}
\value{
List with available collections
}
\description{
This is the basic training endpoint. Given an existing dataset this endpoint will train a model.
}
\examples{
collection <- Collection(name = 'example')
train(collection)
}
