% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\docType{methods}
\name{train,Collection-method}
\alias{train,Collection-method}
\title{This is the basic training endpoint. Given an existing dataset this endpoint will train a model.}
\usage{
\S4method{train}{Collection}(collection_object, version = NULL,
  domain = NULL, api_key = FALSE, ...)
}
\arguments{
\item{version}{for api version}

\item{api_key}{your personal indico API key}

\item{...}{additional arguments to passed to request}

\item{collection}{the collection object for this model}

\item{cloud}{subdomain for indico private cloud}
}
\value{
List with available collections
}
\description{
This is the basic training endpoint. Given an existing dataset this endpoint will train a model.
}
\examples{
collection <- Collection(name='example')
train(collection)
}
